fix: loading single message on refresh during no split view
