+++ src/views/app/detail-panel/message-preview-panel.tsx
@@ -12,6 +12,7 @@ import MailPreview from './preview/mail-preview';
 import PreviewPanelHeader from './preview/preview-panel-header';
 import { EXTRA_WINDOW_ACTION_ID } from '../../../constants';
 import { useAppSelector } from '../../../hooks/redux';
+import { useRequestDebouncedMessage } from '../../../hooks/use-request-debounced-message';
 import { selectMessage } from '../../../store/messages-slice';
 import type { MailsStateType, MessageAction } from '../../../types';
 import { useExtraWindow } from '../extra-windows/use-extra-window';
@@ -37,39 +38,39 @@ export const MessagePreviewPanel: FC<MessagePreviewPanelProps> = ({
 
 	const message = useAppSelector((state: MailsStateType) => selectMessage(state, messageId));
 
+	useRequestDebouncedMessage(messageId, message?.isComplete);
+
 	return (
 		<Container orientation="vertical" mainAlignment="flex-start" crossAlignment="flex-start">
-			{message && (
-				<>
-					{!isInsideExtraWindow && (
-						<PreviewPanelHeader
-							folderId={folderId}
-							itemType={'message'}
-							isRead={message.read}
-							subject={message.subject}
-						/>
-					)}
-					<Container
-						style={{ overflowY: 'auto' }}
-						height="fill"
-						background="gray5"
-						padding={{ horizontal: 'large', bottom: 'small', top: 'large' }}
-						mainAlignment="flex-start"
-					>
-						<Container height="fit" mainAlignment="flex-start" background="gray5">
-							<Padding bottom="medium" width="100%">
-								<MailPreview
-									message={message}
-									expanded
-									isAlone
-									messageActions={actions}
-									isMessageView
-									isInsideExtraWindow={isInsideExtraWindow}
-								/>
-							</Padding>
-						</Container>
+			{!isInsideExtraWindow && (
+				<PreviewPanelHeader
+					folderId={folderId}
+					itemType={'message'}
+					isRead={message?.read}
+					subject={message?.subject}
+				/>
+			)}
+			{message?.isComplete && (
+				<Container
+					style={{ overflowY: 'auto' }}
+					height="fill"
+					background="gray5"
+					padding={{ horizontal: 'large', bottom: 'small', top: 'large' }}
+					mainAlignment="flex-start"
+				>
+					<Container height="fit" mainAlignment="flex-start" background="gray5">
+						<Padding bottom="medium" width="100%">
+							<MailPreview
+								message={message}
+								expanded
+								isAlone
+								messageActions={actions}
+								isMessageView
+								isInsideExtraWindow={isInsideExtraWindow}
+							/>
+						</Padding>
 					</Container>
-				</>
+				</Container>
 			)}
 		</Container>
 	);
